create declarative API for registering a design-system-provider tag name
